requirements.txt
django==2.2.11
djangorestframework
psycopg2
requests